Revert "Updates Packets & Randomizer (#43)"
This reverts commit 179cb505571b278ea3e14454c811c54cb4e06e2a.

+++ Projects/Scripts/Misc/SkillCheck.cs
@@ -224,7 +224,7 @@ namespace Server.Misc
 
         Skills skills = from.Skills;
 
-        if (from.Player && skills.Total / (double)skills.Cap >= Utility.RandomDouble()) //( skills.Total >= skills.Cap )
+        if (from.Player && skills.Total / skills.Cap >= Utility.RandomDouble()) //( skills.Total >= skills.Cap )
           for (int i = 0; i < skills.Length; ++i)
           {
             Skill toLower = skills[i];
